add ft_ping beta

+++ ft_ping/src/main.c
@@ -6,7 +6,7 @@
 /*   By: mbougrin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/18 11:02:44 by mbougrin          #+#    #+#             */
-/*   Updated: 2016/10/26 12:01:07 by mbougrin         ###   ########.fr       */
+/*   Updated: 2016/10/26 12:02:17 by mbougrin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -213,7 +213,7 @@ void					timeout(void)
 void					printSigint(void)
 {
 	t_stc 			*stc = singleton(NULL);
-	double			ret;
+	int				ret;
 	double			one;
 	double 			tmp;
 
@@ -224,7 +224,7 @@ void					printSigint(void)
 		tmp = ret * one;
 		if (tmp >= stc->packetReceiv)
 			break ;
-		ret+= 1;	
+		ret += 1;	
 	}
 	printf("\n--- %s %s statistics ---\n", stc->ip, stc->name);
 	printf("%d packets transmitted, %d received, %f%c packet loss, time %fms\n", \
